requirements.txt
Django==1.8.4
Fabric==1.10.2
GitPython==1.0.1
Werkzeug==0.10.4
arrow==0.6.0
data-importer==2.2.1
django-activity-stream==0.6.0
django-angular==0.7.15
django-bower==5.0.4
django-cors-headers==1.1.0
django-crispy-forms==1.5.1
django-extensions==1.5.5
django-formtools==1.0
django-jsonfield==0.9.15
django-pipeline==1.5.3
django-select2-rocks==0.3
djangorestframework==3.2.3
drf-extensions==0.2.7
ecdsa==0.13
futures==3.0.3
gitdb==0.6.4
jdcal==1.0
openpyxl==2.2.5
paramiko==1.15.2
-e git+https://github.com/rosscdh/pinax-eventlog.git@feature/dateoverride#egg=pinax_eventlog
pycrypto==2.6.1
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
six==1.9.0
smmap==0.9.0
termcolor==1.1.0
wsgiref==0.1.2
xlrd==0.9.4
django-bower==5.0.4